Fixed some formatting issues
Now works OK with my installation of LaTeX (not PDFLaTeX, because of all of the .eps figures). I suspect that the labs that are imported from CLEA (pulsars and hubble) may not look right with other LaTeX installations, because I did a bunch of fiddling to get them placed right on the page. In the future, I'll change all the eps figures to pdfs so that we can use PDFLaTeX, and get things to work right then.

It'd be better to replace those two lab writeups with LaTeX-native versions, instead of importing the Word versions as I'm doing now.