Small edits by TB 28 July 2022
Miscellaneous small edits made by TB just before the fall '22 manual goes to press.

A couple of things worth noting: stellarium1.tex contains some facts about the night sky that are different depending on what time of year the lab is done, so it needs to be edited if you're teaching the course in the spring. Maybe we should automate that?

In the first planetary nebula lab, one of the reproduced pages was misplaced, so it sat on top of the other text. It was also a bit too small, as were some other. I moved it and enlarged it. The other reproduced pages looked small too. That's the sort of thing that might possibly look different in different TeX installations, so I thought I'd mention it. It's pp. 109ff.